# back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Inventory",
        "description": "Documentação da API Inventory gerada automaticamente via Swagger.",
        "contact": {
            "email": "victor.hugo@bizsys.com.br"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/actions": {
            "get": {
                "tags": [
                    "Ações"
                ],
                "summary": "Listar ações",
                "description": "Retorna uma lista paginada de ações.",
                "operationId": "a9d7c561614fd277a97feeccca3fdafb",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página para paginação",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Quantidade de itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao processar sua solicitação. Tente novamente mais tarde."
                    }
                }
            }
        },
        "/api/action": {
            "post": {
                "tags": [
                    "Ações"
                ],
                "summary": "Criar ação",
                "description": "Cria uma nova ação.",
                "operationId": "7c1409027af1e3b942435b2e66a9cbaf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado"
                    },
                    "400": {
                        "description": "Erro de validação"
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao processar sua solicitação. Tente novamente mais tarde."
                    }
                }
            }
        },
        "/api/actions/{id_action}": {
            "get": {
                "tags": [
                    "Ações"
                ],
                "summary": "Exibir ação",
                "description": "Exibe uma ação específica.",
                "operationId": "cdffda7410fffd48a66cb189cbf3ea9e",
                "parameters": [
                    {
                        "name": "id_action",
                        "in": "path",
                        "description": "ID da ação",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    },
                    "404": {
                        "description": "Não encontrado"
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao processar sua solicitação. Tente novamente mais tarde."
                    }
                }
            },
            "put": {
                "tags": [
                    "Ações"
                ],
                "summary": "Atualizar ação",
                "description": "Atualiza uma ação existente.",
                "operationId": "87d1513604beac66ed6e6b1b65809497",
                "parameters": [
                    {
                        "name": "id_action",
                        "in": "path",
                        "description": "ID da ação",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ação atualizada com sucesso"
                    },
                    "400": {
                        "description": "Erro de validação"
                    },
                    "404": {
                        "description": "Não encontrado"
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao processar sua solicitação. Tente novamente mais tarde."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ações"
                ],
                "summary": "Remover ação",
                "description": "Remove uma ação existente.",
                "operationId": "0c33a55388c27d3816de4e2040683dfc",
                "parameters": [
                    {
                        "name": "id_action",
                        "in": "path",
                        "description": "ID da ação",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ação removida com sucesso"
                    },
                    "404": {
                        "description": "Não encontrado"
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao processar sua solicitação. Tente novamente mais tarde."
                    }
                }
            }
        },
...
    },
...
}